+++ libft/ft_memset.c
@@ -6,18 +6,19 @@
 /*   By: maouzal <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/06 22:32:13 by maouzal           #+#    #+#             */
-/*   Updated: 2022/10/07 04:30:15 by maouzal          ###   ########.fr       */
+/*   Updated: 2022/10/13 13:13:22 by maouzal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include <stdlib.h>
 
-void*	ft_memset(void	*b, int	c, int 	len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-	int	x;
+	size_t	x;
 	char	*p;
 
 	p = b;
 	x = 0;
-	while(x < len)
+	while (x < len)
 	{
 		p[x] = c;
 		x++;
